% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditiveG.R
\name{AdditiveG}
\alias{AdditiveG}
\title{Additive Game}
\usage{
AdditiveG(A)
}
\arguments{
\item{A}{Vector with length n.}
}
\value{
The characteristic function of the additive game given by the vector A.
}
\description{
Given a vector, A, it returns the characteristic function of the additive TU game of the vector A in the binary order.
}
\details{
Let A a n-dimensional vector, the characteristic function of the additive TU game is, for each coalition \eqn{S} in \eqn{N}, \eqn{v(S)=\sum_{i\in S}A(i)}{the sum of the elements of the vector that are in S}.
}
\examples{
A=c(0,1,1,2)
AdditiveG(A)
}
\seealso{
\link{SuperAdditiveGame}
}
