% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdmissibleGame.R
\name{AdmissibleGame}
\alias{AdmissibleGame}
\title{Admissible Game}
\usage{
AdmissibleGame(v)
}
\arguments{
\item{v}{Characteristic function in binary order.}
}
\value{
If the TU game given by vector v is  compromise admissible then it returns 1, otherwise it returns 0 and the condition that no satisfies.
}
\description{
Checks if a TU game is compromise admissible.
}
\details{
Let \eqn{(N,v)} a TU game, \eqn{m(N,v)} is the vector of minimal rights and \eqn{M(N,v)} is the vector of utopia payoffs. The game \eqn{(N,v)} is said to be compromise admissible if the core-cover is a not empty set, that is, if it satisfies the conditions:

1) \eqn{m(N,v)\leq M(N,v)}{The minimal rights are smaller than the utopia payoffs}.

2) \eqn{\sum_{i}m_{i}(N,v)\leq \sum_{i}M_i(N,v)}{The sum of the minimal rights is smaller than the sum of the utopia payoffs}.
}
\examples{
v=c(0,0,0,1,1,1,2)
AdmissibleGame(v)
}
\seealso{
\link{CoreSet}, \link{CoreVertices}, \link{UtopiaPayoffs}
}
