% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageMarginal.R
\name{AverageMarginal}
\alias{AverageMarginal}
\title{Average marginal contribution vectors}
\usage{
AverageMarginal(v)
}
\arguments{
\item{v}{Characteristic function in binary order.}
}
\value{
Matrix with the average marginal contribution vectors.
}
\description{
This function returns the average marginal contribution vectors of a TU game \eqn{(N,v)}.
}
\details{
Given a TU game \eqn{(N,v)}, we define the average marginal contribution vectors for each coalition \eqn{S} in \eqn{N} by,
\deqn{A(S)=\sum_{k\in S}\frac{1}{|S|}(v(S)-v(S\backslash \{k\}))}{A(S)=\sum(v(S)-v(S\ {k})),k in S}
}
\examples{
v=c(0,0,0,1,1,1,2)
AverageMarginal(v)
}
\references{
Nowark,AS. & Radzik,T. (1994). A Solidarity Value for n-Person Transferable Utility Games. International Journal of Game Theory, vol. 23; 43-48.
}
\seealso{
\link{Marginal}, \link{Solidarity}
}
