% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basebin.R
\name{Basebin}
\alias{Basebin}
\title{Binary base}
\usage{
Basebin(n)
}
\arguments{
\item{n}{Number of players.}
}
\value{
A matrix \eqn{n \times (2^n-1)}{n x 2^n-1}.
}
\description{
It provides a matrix which is the basis of a TU game in the binary order.
}
\details{
The basis of a TU game in the binary order is a matrix with \eqn{2^n-1} rows and \eqn{n} columns where
we order the coalitions of a TU game with the binary order and, in each row we
assign 1 if the player is in the coalition and 0 otherwise.
}
\examples{
n=5
Basebin(n)
}
\seealso{
\link{Baselex}, \link{Bin2lex}, \link{Lex2bin}
}
