% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bin2lex.R
\name{Bin2lex}
\alias{Bin2lex}
\title{Change from binary to lexicographical}
\usage{
Bin2lex(vold)
}
\arguments{
\item{vold}{Characteristic function in binary order.}
}
\value{
The characteristic function in lexicographic order.
}
\description{
It changes the characteristic function of a TU game with binary order to the lexicographic order.
}
\examples{
vold=seq(1:31)
Bin2lex(vold)
}
\seealso{
\link{Basebin}, \link{Baselex}, \link{Lex2bin}
}
