% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charasteristicdata.R
\name{Characteristicdata}
\alias{Characteristicdata}
\title{Characteristic data}
\usage{
Characteristicdata(v)
}
\arguments{
\item{v}{A characteristic function of a TU game.}
}
\value{
n. Number of players.

nC. Number of coalitions.
}
\description{
Given a characteristic function of a TU game, it provides his number of players and the number of coalitions.
}
\examples{
v=seq(1:31)
Characteristicdata(v)
}
