% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coalitionnumber.R
\name{Coalitionnumber}
\alias{Coalitionnumber}
\title{Coalition number}
\usage{
Coalitionnumber(S, n)
}
\arguments{
\item{S}{Coalition.}

\item{n}{Number of players.}
}
\value{
Position of the coalition \eqn{S} in the binary order.
}
\description{
It gives the position of a coalition in a TU game given by the binary order.
}
\examples{
S=c(1,2,3)
n=3
Coalitionnumber(S,n)
}
\seealso{
\link{Getplayers}
}
