% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexGame.R
\name{ConvexGame}
\alias{ConvexGame}
\title{Convex Game}
\usage{
ConvexGame(v)
}
\arguments{
\item{v}{Characteristic function in binary order.}
}
\value{
If the TU game given by vector \eqn{v} is convex, then the function returns 1, otherwise it returns 0 and it provides the coalitions which don't satisfy the convex conditions.
}
\description{
Checks if a TU game is convex.
}
\details{
A game \eqn{(N,v)} is convex if \eqn{v(S \cap T) + v(S \cup T) \ge v(S)+v(T)}{v(S v T) + v(S ^ T) \ge v(S)+v(T)}, for all
\eqn{S,T} in \eqn{N}.
}
\examples{
v=c(0,0,0,1,1,1,2)
ConvexGame(v)
}
