% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corecenter.R
\name{Corecenter}
\alias{Corecenter}
\title{Core-center}
\usage{
Corecenter(v, name = FALSE)
}
\arguments{
\item{v}{Characteristic function in binary order.}

\item{name}{A logical value.}
}
\value{
It returns the core-center of a TU game given by
vector \eqn{v}.
}
\description{
The core-center of a TU balanced game.
}
\details{
The core of a TU game \eqn{(N,v)} is the set of those imputations \eqn{x} in \eqn{R^n}, \eqn{x_1+...+x_n=v(N)}{x1+...+xn=v(N)},
 such that \eqn{x(S)\ge v(S)} for all coalition \eqn{S} in \eqn{N}, where,
 \eqn{x(S)=\sum_{i\in S} x_i}{x(S)=\sum xi, i in S}.

 \deqn{C(N,v)=\{x\in I(N,v):x(S)\ge v(S)\; \forall S\in N\}}{C(N,v)={x in I(N,v):x(S)\ge v(S) for all S in N}}

The core-center (CC) is defined as the centroid (center of gravity) of
the core, so it can only be computed if the game is balanced. It is the expectation of the uniform distribution over the
core of the game. Let \eqn{\mu} be the
\eqn{(n-1)}-dimensional Lebesgue measure and denote \eqn{V(C)=\mu(C(N,v))} the volume (measure) of the core. If \eqn{V(C)>0} then for each \eqn{i} in \eqn{N},
\deqn{CC_i(N,v)=\frac{1}{V(C)}\int_{C(N,v)}x_i d\mu}{1/V(C)*[S xd\mu], where S is the integral and it is taken over C(v).}
}
\examples{
v=c(0,0,0,1,1,1,2)
Corecenter(v)
}
\references{
Gonzalez-Díaz, J., & Sánchez-Rodríguez, E. (2007). A natural selection from the core of a TU game: the core-center. International Journal of Game Theory, 36(1), 27-46.
}
\seealso{
\link{BalancedGame}, \link{CoreSet}, \link{CoreVertices}, \link{PlotSolution}
}
