% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreCoverSet.R
\name{CoreCoverSet}
\alias{CoreCoverSet}
\title{Core cover set}
\usage{
CoreCoverSet(v, set = TRUE, col = NULL)
}
\arguments{
\item{v}{Characteristic function in binary order:

For 2 players the characteristic function must be introduced with vector \eqn{v=(v_{1},v_{2},v_{12})}{v=(v{1},v{2},v{12})}

For 3 players the characteristic function must be introduced with vector
 \eqn{v=(v_{1},v_{2},v_{12},v_{3},v_{13},v_{23},v_{123})}{v=(v{1},v{2},v{12},v{3},v{13},v{23},v{123})}.

 For 4 players the characteristic function must be introduced with vector

\eqn{v=(v_{1},v_{2},v_{12},v_{3},v_{13},v_{23},v_{123},v_{4},v_{14},v_{24},v_{124},v_{34},v_{134},v_{234},v_{1234})}{v=(v{1},v{2},v{12},v{3},v{13},v{23},v{123},v{4},v{14},v{24},v{124},v{34},v{134},v{234},v{1234})}.}

\item{set}{A logical value. By default, set=TRUE.}

\item{col}{Color.}
}
\value{
If set=TRUE, it displays the core-cover of the game defined by \eqn{v} inside the imputation set.
If set=FALSE, it displays the core-cover inside the last graph.
}
\description{
Draws the core-cover of a 2, 3 or 4 person compromise admissible TU game.
}
\details{
Let \eqn{(N,v)} a TU game and let \eqn{m(N,v)} the vector of minimal rights and \eqn{M(N,v)} the vector of the utopia payoffs, the core-cover is the set of all the imputations that assure each player his minimum
rights and no more than his utopia payment.
\deqn{CCover(N,v)=\{x\in R^n: m(N,v)\le x\le M(N,v)\}}{CC(N,v)={x: m(N,v)\le x\le M(N,v)}}
}
\examples{
v=c(0,0,0,0,0,0,0,0,1,4,1,3,6,8,10)
CoreCoverSet(v)
}
\references{
Tijs, S. H. (1981). Nash equilibria for noncooperative n-person games in normal form. Siam Review, 23(2), 225-237.
}
\seealso{
\link{CoreCoverVertices}, \link{CoreSet}, \link{EpsilonCoreSet}, \link{HarsanyiSet}, \link{ImputationSet}, \link{LeastCoreSet}, \link{UtopiaPayoffs}, \link{WeberSet}
}
