% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreCoverVertices.R
\name{CoreCoverVertices}
\alias{CoreCoverVertices}
\title{Core cover vertices}
\usage{
CoreCoverVertices(v)
}
\arguments{
\item{v}{Characteristic function in the binary order.}
}
\value{
The vertices of the core cover.
}
\description{
This function returns the vertices of the core cover of the TU game.
}
\details{
Let \eqn{(N,v)} a TU game and let \eqn{m(N,v)} the vector of minimal rights and \eqn{M(N,v)} the vector of the utopia payoffs, the core-cover is the set of all the imputations that assure each player his minimum
rights and no more than his utopia payment.
\deqn{CCover(N,v)=\{x\in R^n: m(N,v)\le x\le M(N,v)\}}{CC(N,v)={x: m(N,v)\le x\le M(N,v)}}
}
\examples{
v=c(0,0,0,0,0,0,0,0,1,4,1,3,6,8,10)
CoreCoverVertices(v)
}
\references{
Tijs, S. H. (1981). Nash equilibria for noncooperative n-person games in normal form. Siam Review, 23(2), 225-237.
}
\seealso{
\link{CoreCoverSet}, \link{CoreVertices}, \link{EpsilonCoreVertices}, \link{ImputationVertices}, \link{LeastCoreVertices}, \link{UtopiaPayoffs}, \link{WeberVertices}
}
