% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet.R
\name{CoreSet}
\alias{CoreSet}
\title{Core set}
\usage{
CoreSet(v, set = TRUE, col = NULL)
}
\arguments{
\item{v}{Characteristic function in binary order:

For 2 players the characteristic function must be introduced with vector \eqn{v=(v_{1},v_{2},v_{12})}{v=(v{1},v{2},v{12})}

For 3 players the characteristic function must be introduced with vector
 \eqn{v=(v_{1},v_{2},v_{12},v_{3},v_{13},v_{23},v_{123})}{v=(v{1},v{2},v{12},v{3},v{13},v{23},v{123})}.

 For 4 players the characteristic function must be introduced with vector

\eqn{v=(v_{1},v_{2},v_{12},v_{3},v_{13},v_{23},v_{123},v_{4},v_{14},v_{24},v_{124},v_{34},v_{134},v_{234},v_{1234})}{v=(v{1},v{2},v{12},v{3},v{13},v{23},v{123},v{4},v{14},v{24},v{124},v{34},v{134},v{234},v{1234})}.}

\item{set}{A logical value. By default, set=TRUE.}

\item{col}{Color.}
}
\value{
If set=TRUE, it displays the core of the game defined by \eqn{v} inside the imputation set.
If set=FALSE, it displays the core inside the last graph.
}
\description{
Draws the core of a 2, 3 or 4 person compromise admissible TU game.
}
\details{
The core of a TU game \eqn{(N,v)} is the set of those imputations \eqn{x} in \eqn{R^n}, \eqn{x_1+...+x_n=v(N)}{x1+...+xn=v(N)},
 such that \eqn{x(S)\ge v(S)} for all coalition \eqn{S} in \eqn{N}, where,
 \eqn{x(S)=\sum_{i\in S} x_i}{x(S)=\sum xi, i in S}.

 \deqn{C(N,v)=\{x\in I(N,v):x(S)\ge v(S)\; \forall S\in N\}}{C(N,v)={x in I(N,v):x(S)\ge v(S) for all S in N}}

The core allocations provide the agents with an incentive to maintain the grand coalition and it can be empty.
}
\examples{
v=c(0,0,0,0,0,0,0,0,1,4,1,3,6,8,10)
CoreSet(v)
}
\references{
Edgeworth, F. Y. (1881). Mathematical psychics: An essay on the application of mathematics to the moral sciences (No. 10). CK Paul.

Gillies, D. B. (1953). Some theorems on n-person games. Princeton University.
}
\seealso{
\link{BalancedGame}, \link{CoreCoverSet}, \link{CoreVertices}, \link{EpsilonCoreSet}, \link{HarsanyiSet}, \link{ImputationSet}, \link{LeastCoreSet}, \link{WeberSet}
}
