% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativeAwardsCurve.R
\name{CumulativeAwardsCurve}
\alias{CumulativeAwardsCurve}
\title{Cumulative awards curve}
\usage{
CumulativeAwardsCurve(
  v,
  Solutions,
  Solution = Corecenter,
  col = NULL,
  legend = TRUE
)
}
\arguments{
\item{v}{Characteristic function in binary order.}

\item{Solutions}{The solutions: Alexia, Corecenter, DuttaRay, Nucleolus, PerCapitaNucleolus, PreNucleolus, Shapley, Solidarity and Tau.}

\item{Solution}{Principal solution: Alexia, Corecenter, DuttaRay, Nucleolus, PerCapitaNucleolus, PreNucleolus, Shapley, Solidarity and Tau. By default, Solution=Corecenter.}

\item{col}{The colors. If col = NULL then the sequence of default colors is:
c("red", "blue", "green", "yellow", "pink", "coral4", "darkgray", "burlywood3", "black", "darkorange", "darkviolet").}

\item{legend}{A logical value. The color legend is shown if legend = TRUE.}
}
\value{
The graphical representation of the cumulative awards curves of a puntual solution (or several solutions) for a TU game.
}
\description{
This function returns the cumulative awards curve for a TU game.
}
\details{
Given a TU game \eqn{(N,v)}, the cumulative awards curve of a solution \eqn{S'} with respect of a solution \eqn{S} for the game is the polygonal path connecting the \eqn{n+1} points
\deqn{(0,0), (\frac{S'_1(N,v)}{v(N)},\frac{S_1(N,v)}{v(N)}),\dots,(\frac{\sum_{i=1}^{n-1}S'_i(N,v)}{v(N)},\frac{\sum_{i=1}^{n-1}S_i(N,v)}{v(N)}),(1,1).}{%
(0,0) , (S'1(N,v)/v(N),S1(N,v)/v(N)) , ((S'1(N,v)+S'2(N,v))/v(N) , ((S1(N,v)+S2(N,v))/v(N) ,\dots , (1,1).}

The cumulative claims-awards curve fully captures the Lorenz ranking of solutions:
if a solution \eqn{S} Lorenz-dominates a rule \eqn{S'} then, for each TU game, the cumulative awards  curve of \eqn{S}
lies above the cumulative curve of \eqn{S'}.
}
\examples{
v=c(0,0,0,1,1,1,20)
Solution=Corecenter
Solutions=c(Shapley,Nucleolus)
CumulativeAwardsCurve(v,Solutions,Solution)
}
\references{
Lorenz, M. O. (1905). Methods of measuring the concentration of wealth. Publications of the American statistical association, 9(70), 209-219.

Mirás Calvo, M. A., Núñez Lugilde, I., Quinteiro Sandomingo, C., and Sánchez Rodríguez, E. (2020). Deviation from proportionality and Lorenz-domination for claims problems. ECOBAS.
}
\seealso{
\link{Alexia}, \link{Corecenter}, \link{DeviationIndex}, \link{DuttaRay}, \link{LorenzDominance}, \link{Nucleolus}, \link{PerCapitaNucleolus}, \link{PreNucleolus}, \link{Shapley}, \link{Solidarity}, \link{Tau}
}
