% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeviationIndex.R
\name{DeviationIndex}
\alias{DeviationIndex}
\title{Deviation index}
\usage{
DeviationIndex(v, S, R)
}
\arguments{
\item{v}{Characteristic function in binary order.}

\item{S}{A solution: Alexia, Corecenter, DuttaRay, Nucleolus, PerCapitaNucleolus, PreNucleolus, Shapley, Solidarity and Tau.}

\item{R}{A solution: Alexia, Corecenter, DuttaRay, Nucleolus, PerCapitaNucleolus, PreNucleolus, Shapley, Solidarity and Tau.}
}
\value{
The  deviation index and the signed deviation index of a solution for a TU game.
}
\description{
This function returns the general deviation index and the general signed deviation index of a rule for a claims problem.
}
\details{
Given a TU game \eqn{(N,v)}, the cumulative awards curve of a solution \eqn{S} with respect of a solution \eqn{R} for the game is the polygonal path connecting the \eqn{n+1} points
\deqn{(0,0), (\frac{S_1(N,v)}{v(N)},\frac{R_1(N,v)}{v(N)}),\dots,(\frac{\sum_{i=1}^{n-1}S_i(N,v)}{v(N)},\frac{\sum_{i=1}^{n-1}R_i(N,v)}{v(N)}),(1,1).}{%
(0,0) , (S1(N,v)/v(N),R1(N,v)/v(N)) , ((S1(N,v)+S2(N,v))/v(N) , ((R1(N,v)+R2(N,v))/v(N) ,\dots , (1,1).}
The signed deviation index of the solution \eqn{S} with respect to the rule \eqn{R} for the TU game \eqn{(N,v)}, denoted by \eqn{I(S(N,v),R(N,v))}, is
the ratio of the area that lies between the identity line and the cumulative curve over the total area under the identity line.

Let \eqn{R_0(N,v)=0}{R0(N,v)=0} and \eqn{S_0(N,v)=0}{S0(N,v)=0}. For each \eqn{k=1,\dots,n} define,
\eqn{X_k=\frac{1}{v(N)} \sum_{j=0}^{k} R_j}{Xk=(R0+\dots+Rk)/v(N)} and
\eqn{Y_k=\frac{1}{v(N)} \sum_{j=0}^{k} S_j}{Xk=(S0+\dots+Sk)/v(N)}. Then,
\deqn{I(S(N,v),R(N,v))=1-\sum_{k=1}^{n}(X_{k}-X_{k-1})(Y_{k}+Y_{k-1}).}{I(R(N,v),S(N,v))=1-\sum (Xk-X(k-1))(Yk+Y(k-1)) where the sum goes from k=1 to n.}
 In general  \eqn{-1 \le I(S(N,v),R(N,v)) \le 1}.

The deviation index of the solution \eqn{R} with respect to the solution \eqn{S} for the game \eqn{(N,v)}, denoted by \eqn{I^{+}(R(N,v),S(N,v))}{I+(R(N,v),S(N,v))}, is
the ratio of the area between the line of the cumulative sum of solution \eqn{R} and the cumulative curve over the area under the line \eqn{x=y}.
In general,  \eqn{0 \le I^{+}(S(N,v),R(N,v)) \le 1}{0 \le I+(S(N,v),R(N,v)) \le 1}.
}
\examples{
v=c(0,0,0,1,1,1,2)
S=Corecenter
R=Shapley
DeviationIndex(v,S,R)
}
\references{
Ceriani, L. and Verme, P. (2012). The origins of the Gini index: extracts from Variabilitá e Mutabilitá (1912) by Corrado Gini. The Journal of Economic Inequality, 10(3), 421-443.

Mirás Calvo, M. A., Núñez Lugilde, I., Quinteiro Sandomingo, C., and Sánchez Rodríguez, E. (2020). Deviation from proportionality and Lorenz-domination for claims problems. ECOBAS.
}
