% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualGame.R
\name{DualGame}
\alias{DualGame}
\title{Dual Game}
\usage{
DualGame(v)
}
\arguments{
\item{v}{Characteristic function in binary order.}
}
\value{
The characteristic function of the dual game of \eqn{(N,v)}.
}
\description{
It returns the characteristic function of the dual game of a TU game.
}
\details{
Given a TU game \eqn{(N,v)}, its dual game, \eqn{(N,v^c)}{(N,vc)}, is defined, for all \eqn{S} in \eqn{N} by, \deqn{v^c(S)=v(N)-v(N\backslash S)}{vc(S)=v(N)-v(N\S)}
}
\examples{
v=seq(1:31)
DualGame(v)
}
